+++ BaiShengGuangDianManager/ApiManagement/Models/PlanManagementModel/ProductionPlan.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using ApiManagement.Models.BaseModel;
+using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
 
 namespace ApiManagement.Models.PlanManagementModel
 {
@@ -13,9 +14,9 @@ namespace ApiManagement.Models.PlanManagementModel
     }
     public class ProductionPlanDetail : ProductionPlan
     {
-        public int PlannedConsumption { get; set; }
-        public int ActualConsumption { get; set; }
-        public int ExtraConsumption { get; set; }
+        public decimal PlannedConsumption { get; set; }
+        public decimal ActualConsumption { get; set; }
+        public decimal ExtraConsumption { get; set; }
         public decimal PlannedCost { get; set; }
         public decimal ActualCost { get; set; }
         public List<ProductionPlanBillStockDetail> FirstBill  = new List<ProductionPlanBillStockDetail>();
